+++ tangy-form-response-model.js
@@ -5,7 +5,6 @@ export class TangyFormResponseModel {
     // Placeholders for where element.getProps() info will go.
     this.form = {}
     this.items = []
-    this.inputs = []
     // States.
     this.complete = false
     // Focus indexes.
@@ -17,8 +16,39 @@ export class TangyFormResponseModel {
     this.startDatetime = (new Date()).toLocaleString(),
     this.startUnixtime = Date.now(),
     this.uploadDatetime = ''
+    if (props && props.hasOwnProperty('inputs')) delete props.inputs
     Object.assign(this, props)
   }
+
+  get inputs() {
+    // Reduce to an array.
+    return this.items.reduce((inputsArray, item) => {
+      item.inputs.forEach(input => {
+        if (input.tagName === 'TANGY-CARDS') {
+          input.value.forEach(card => card.value.forEach(input => inputsArray.push(input)))
+        } else {
+          inputsArray.push(input)
+        }
+      })
+      return inputsArray
+    }, [])
+  }
+  get inputsByName() {
+    // Reduce to an object keyed on input.name. If multiple inputs with the same name, put them in an array.
+    return this.inputs.reduce((inputsObject, input) => {
+      if (inputsObject.hasOwnProperty(input.name)) {
+        if (Array.isArray(inputsObject[input.name])) {
+          inputsObject[input.name] = inputsObject[input.name].push(input)
+        } else {
+          inputsObject[input.name] = [input, inputsObject[input.name]]
+        }
+      } else {
+        inputsObject[input.name] = input
+      }
+      return inputsObject
+    }, {})
+  }
+
 }
 
 function uuid() {